refactor large index.js file into separate modules for config, routes, controllers, and services
