+++ server/REFACTORING_SUMMARY.md
@@ -1,0 +1,113 @@
+# Server Refactoring Summary
+
+## Overview
+Successfully refactored the monolithic `index.js` file into a clean modular architecture following MVC pattern with separate config, routes, controllers, and services.
+
+## New Directory Structure
+
+```
+server/
+├── index.js (main entry point - now only ~60 lines)
+├── messages.js (MongoDB models)
+├── package.json
+├── config/
+│   ├── cloudinary.js (Cloudinary configuration)
+│   └── multer.js (File upload middleware configuration)
+├── controllers/
+│   ├── messagesController.js (Message business logic)
+│   ├── imageController.js (Image upload/fetch/delete logic)
+│   └── videoController.js (Video upload/fetch/delete logic)
+├── routes/
+│   ├── messageRoutes.js (Message endpoints)
+│   ├── imageRoutes.js (Image endpoints)
+│   └── videoRoutes.js (Video endpoints)
+└── services/
+    ├── messageService.js (Message database operations)
+    ├── imageService.js (Image database & Cloudinary operations)
+    └── videoService.js (Video database & Cloudinary operations)
+```
+
+## Module Breakdown
+
+### Config (`config/`)
+- **cloudinary.js**: Cloudinary configuration using environment variables
+- **multer.js**: File upload middleware with separate configs for images (10MB) and videos (100MB)
+
+### Routes (`routes/`)
+All routes follow RESTful conventions:
+
+#### Message Routes (`/api/message`)
+- `GET /` - Get all messages
+- `POST /` - Create new message
+
+#### Image Routes (`/api/images`)
+- `GET /` - Get all images
+- `GET /:category` - Get images by category
+- `POST /upload` - Upload single image
+- `POST /upload-multiple` - Upload multiple images
+- `DELETE /:id` - Delete image
+
+#### Video Routes (`/api/videos`)
+- `GET /` - Get all videos
+- `GET /:category` - Get videos by category
+- `POST /upload` - Upload single video
+- `POST /upload-multiple` - Upload multiple videos
+- `DELETE /:id` - Delete video
+
+### Controllers (`controllers/`)
+Handle request/response logic and validation:
+
+- **messagesController.js**: `getMessages`, `postMessage`
+- **imageController.js**: `getImages`, `getImagesByCategoryHandler`, `uploadSingleImage`, `uploadMultipleImages`, `deleteImageHandler`
+- **videoController.js**: `getVideos`, `getVideosByCategory`, `uploadSingleVideo`, `uploadMultipleVideos`, `deleteVideoHandler`
+
+### Services (`services/`)
+Handle database operations and external API calls:
+
+- **messageService.js**: `getAllMessages`, `createMessage`
+- **imageService.js**: `getAllImages`, `getImagesByCategory`, `uploadImageToCloudinary`, `saveImageMetadata`, `deleteImage`
+- **videoService.js**: `getAllVideos`, `fetchVideosByCategory`, `uploadVideoToCloudinary`, `saveVideoMetadata`, `deleteVideo`
+
+## Key Features
+
+### Backward Compatibility
+The refactored code maintains backward compatibility with old endpoints:
+- `/api/upload` → `/api/images/upload`
+- `/api/upload-multiple` → `/api/images/upload-multiple`
+- `/api/upload-video` → `/api/videos/upload`
+- `/api/upload-multiple-videos` → `/api/videos/upload-multiple`
+
+### Separation of Concerns
+- **Config**: Centralized configuration management
+- **Routes**: Clean endpoint definitions with middleware
+- **Controllers**: Request validation and response formatting
+- **Services**: Reusable business logic and data operations
+
+### Benefits
+1. **Maintainability**: Each module has a single responsibility
+2. **Testability**: Services and controllers can be unit tested independently
+3. **Scalability**: Easy to add new features without modifying existing code
+4. **Reusability**: Service functions can be used across multiple controllers
+5. **Readability**: Clear separation makes code easier to understand
+
+## Files Modified
+- ✅ `index.js` - Reduced from 728 lines to ~60 lines
+- ✅ Created `config/cloudinary.js`
+- ✅ Created `config/multer.js`
+- ✅ Created `services/messageService.js`
+- ✅ Created `services/imageService.js`
+- ✅ Updated `services/videoService.js`
+- ✅ Created `controllers/messagesController.js`
+- ✅ Created `controllers/imageController.js`
+- ✅ Updated `controllers/videoController.js`
+- ✅ Created `routes/messageRoutes.js`
+- ✅ Created `routes/imageRoutes.js`
+- ✅ Updated `routes/videoRoutes.js`
+
+## Next Steps (Optional)
+1. Add input validation middleware (e.g., express-validator)
+2. Add authentication/authorization middleware
+3. Create error handling middleware
+4. Add API documentation (e.g., Swagger)
+5. Add unit tests for services and controllers
+6. Add logging middleware (e.g., morgan, winston)
