+++ server/config/cloudinary.js
@@ -1,0 +1,11 @@
+const cloudinary = require("cloudinary").v2;
+
+// Configure Cloudinary with env variables
+cloudinary.config({
+  cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
+  api_key: process.env.CLOUDINARY_API_KEY,
+  api_secret: process.env.CLOUDINARY_API_SECRET,
+  secure: true,
+});
+
+module.exports = cloudinary;
